11840: combine error handling into one class for txn cntrlr and loan update job
