+++ app/controllers/admin/accounting/transactions_controller.rb
@@ -79,7 +79,8 @@ class Admin::Accounting::TransactionsController < Admin::AdminController
     # We use it because we want to rollback transaction creation or updates if there are errors.
     ActiveRecord::Base.transaction do
       if @transaction.save
-        if (error_msg = handle_qb_errors { run_updater(project: @transaction.project) })
+        loan = @transaction.loan
+        if (error_msg = Accounting::QB::ErrorHandler.new(loan: loan).handle_qb_errors { run_updater(project: loan) })
           @transaction.errors.add(:base, error_msg)
           raise ActiveRecord::Rollback
         end
@@ -89,7 +90,7 @@ class Admin::Accounting::TransactionsController < Admin::AdminController
 
   def sync_and_handle_errors
     ActiveRecord::Base.transaction do
-      if (error_msg = handle_qb_errors { run_updater(project: @loan) })
+      if (error_msg = Accounting::QB::ErrorHandler.new(loan:@loan).handle_qb_errors { run_updater(project: @loan) })
         flash[:error] = error_msg
         raise ActiveRecord::Rollback
       end
@@ -119,58 +120,6 @@ class Admin::Accounting::TransactionsController < Admin::AdminController
     @accounts = Accounting::Account.asset_accounts - Division.root.accounts
     @customers = Accounting::Customer.all.order(:name)
     @vendors = Accounting::QB::Vendor.all.order(:name)
-  end
-
-  # Runs the given block and handles any Quickbooks errors.
-  # Returns the error message (potentially with HTML) as a string if there was an error, else returns nil.
-  # Notifies admins if error is not part of normal operation.
-  # This is only used for creating new transactions
-  def handle_qb_errors
-    begin
-      yield
-    rescue Accounting::QB::DataResetRequiredError => e
-      Rails.logger.error e
-      error_msg = t("quickbooks.data_reset_required", settings: settings_link).html_safe
-    rescue Accounting::QB::UnprocessableAccountError => e
-      # Duplicated in QuickbooksUpdateJob, should be DRYed up in refactor.
-      Accounting::SyncIssue.create!(loan: e.loan, accounting_transaction: e.transaction,
-                                    message: :unprocessable_account, level: :error, custom_data: {})
-    rescue Accounting::QB::IntuitRequestError => e
-      txn = e.transaction
-      is_matching_error = e.message.include?("matched to a downloaded transaction")
-      intuit_error_type = is_matching_error ? :matched_transaction : :unknown_intuit_request_exception
-      Accounting::SyncIssue.create!(loan: txn.loan,
-                                    accounting_transaction: txn,
-                                    message: intuit_error_type,
-                                    level: :warning, # want to still see ledger
-                                    custom_data: {date: txn.txn_date, qb_id: txn.qb_id})
-      changes_hash = {txn_changes: txn.previous_changes,
-                      line_item_changes: txn.line_items.map { |li| li.previous_changes }}
-      Rails.logger.error("Accounting::QB::IntuitRequestError #{intuit_error_type} #{e.message}. Txn date: #{txn.txn_date}. Txn qb id: #{txn.qb_id}. Changes: #{changes_hash}")
-      error_msg = t("quickbooks.#{intuit_error_type}",
-                    date: txn.txn_date,
-                    qb_id: txn.qb_id)
-    rescue Quickbooks::ServiceUnavailable => e
-      Rails.logger.error e
-      error_msg = t("quickbooks.service_unavailable")
-    rescue Quickbooks::MissingRealmError,
-           Accounting::QB::NotConnectedError,
-           Quickbooks::AuthorizationFailure => e
-      Rails.logger.error e
-      error_msg = t("quickbooks.authorization_failure", settings: settings_link).html_safe
-    rescue Quickbooks::InvalidModelException,
-           Quickbooks::Forbidden,
-           Quickbooks::ThrottleExceeded,
-           Quickbooks::TooManyRequests,
-           Quickbooks::IntuitRequestException => e
-      Rails.logger.error e
-      ExceptionNotifier.notify_exception(e)
-      error_msg = t("quickbooks.misc", msg: e)
-    rescue Accounting::QB::NegativeBalanceError => e
-      Rails.logger.error e
-      error_msg = t("quickbooks.negative_balance", amt: e.prev_balance)
-    end
-    error_msg
   end
 
   def run_updater(project:)
