+++ app/jobs/quickbooks_update_job.rb
@@ -11,40 +11,9 @@ class QuickbooksUpdateJob < QuickbooksJob
     task.set_activity_message("syncing_with_quickbooks")
     loans.each_with_index do |loan, index|
       task.set_activity_message("updating_loans", so_far: (index), total: loans.count)
-      begin
-        updater.update_loan(loan)
-      rescue Accounting::QB::UnprocessableAccountError => e
-        Accounting::SyncIssue.create!(loan: e.loan, accounting_transaction: e.transaction,
-                                      message: :unprocessable_account, level: :error, custom_data: {})
-      rescue Quickbooks::ServiceUnavailable
-        # This is an error b/c we may have been in the middle of creating interest txns
-        Accounting::SyncIssue.create!(level: :error, loan: loan, message: :quickbooks_unavailable_recalc)
-        raise # If QB is down, no point in continuing, so re-raise
-      rescue Accounting::QB::IntuitRequestError => e
-        txn = e.transaction
-        is_matching_error = e.message.include?("matched to a downloaded transaction")
-        intuit_error_type = is_matching_error ? :matched_transaction : :unknown_intuit_request_exception
-        Accounting::SyncIssue.create!(loan: txn.loan,
-                                      accounting_transaction: txn,
-                                      message: intuit_error_type,
-                                      level: :warning, # want to still see ledger
-                                      custom_data: {date: txn.txn_date, qb_id: txn.qb_id})
-        changes_hash = {txn_changes: txn.previous_changes,
-                        line_item_changes: txn.line_items.map { |li| li.previous_changes }}
-        Rails.logger.error("Accounting::QB::IntuitRequestError #{intuit_error_type} #{e.message}. Txn date: #{txn.txn_date}. Txn qb id: #{txn.qb_id}. Changes: #{changes_hash}")
-      rescue StandardError => e
-        # If there is an unhandled error updating an individual loan, we don't want the whole process to fail.
-        # We let the user know that there was a system error and we've been notified.
-        # But we don't expose the original error message to the user since it won't be intelligble
-        # and could be a security issue.
-        Accounting::SyncIssue.create!(level: :error, loan: loan, message: :system_error_recalc)
-
-        # We want to receive a loud notification about an unhandled error.
-        # If we find this is often generating a lot of similar errors
-        # then we should really start using Sentry or some other service to group them.
-        notify_of_error(e, data: {context: "Unhandled error during loan update", loan_id: loan.id})
-      end
+      Accounting::QB::ErrorHandler.new(loan:loan, in_background_job: true).handle_qb_errors { updater.update_loan(loan) }
     end
+
     updater.qb_connection.update_last_updated_at(started_update_at)
 
     # Even if there have been per-loan errors, we still consider the task completed.
